constellation tracing of maybe_close_random_pipeline
Since maybe_close_random_pipeline is called in the core
loop of constellation, tracing the function if nothing
happens and it immediately returns is pointless and
just wastes performance (when tracing).

Instead, we can just trace after an early return.
Since generating a random number is presumably also not that heavy,
it's sufficient to trace an event when we are actually
closing a pipeline.

To improve readability of the manual instrumentation,
add a macro, that wraps the tracing macro and adds our feature flags on
top.
Also re-export tracing as servo_tracing, so that the
macro can access it, without the user crate needing
to depend on tracing.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>